Colocando a lógica da interface em um arquivo separado
